(refactor) model/data
