+++ src/main/java/com/grepp/moodlink/app/model/data/book/BookRepository.java
@@ -1,6 +1,6 @@
-package com.grepp.moodlink.app.model.book;
+package com.grepp.moodlink.app.model.data.book;
 
-import com.grepp.moodlink.app.model.book.entity.Book;
+import com.grepp.moodlink.app.model.data.book.entity.Book;
 import org.springframework.data.repository.CrudRepository;
 
 public interface BookRepository extends CrudRepository<Book, Long> {
