+++ src/main/java/com/grepp/moodlink/app/model/data/music/MusicRepository.java
@@ -1,6 +1,6 @@
-package com.grepp.moodlink.app.model.music;
+package com.grepp.moodlink.app.model.data.music;
 
-import com.grepp.moodlink.app.model.music.entity.Music;
+import com.grepp.moodlink.app.model.data.music.entity.Music;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
